tehtävä tehty

+++ moduuli2/tehtava2.5.py
@@ -3,9 +3,9 @@ naulat = float(input("Anna naulat: "))
 luodit = float(input("Anna luodit: "))
 
 #muunnetaan grammoiksi, 1 leviksa vastaa 20 naulaa, 1 naula 32 luotia ja 1 luoti 13.3 grammaa.
-leiviskatG = float(leiviskat*20*32*13.3)
-naulatG = float(naulat*32*13.3)
-luoditG = float(luodit*13.3)
+leiviskatG = leiviskat*20*32*13.3
+naulatG = naulat*32*13.3
+luoditG = luodit*13.3
 
 kilot = int(((leiviskatG+naulatG+luoditG)/1000))
 grammat = float(leiviskatG+naulatG+luoditG-(kilot*1000))
